[ir] Update `Value` parameters.
This CL updates the various instructions to allow the provided `Value`s
to be `nullptr`. These will be checked by the validator to ensure they
do not move through to the backends.

The `AddOperand` and `AddOperands` methods are updated to take the
operand position. The provided position is then validated against
the operand list to validate that it is the next appended item. This
verifies that the operands are added at the expected indices so any call
to `SetOperand` will work as expected.

Bug: tint:1718
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/137360
Commit-Queue: Ben Clayton <bclayton@google.com>
Kokoro: Kokoro <noreply+kokoro@google.com>

+++ src/tint/ir/instruction.h
@@ -53,7 +53,7 @@ class Instruction : public utils::Castable<Instruction, Value> {
     /// Set an operand at a given index.
     /// @param index the operand index
     /// @param value the value to use
-    virtual void SetOperand(uint32_t index, ir::Value* value) = 0;
+    virtual void SetOperand(size_t index, ir::Value* value) = 0;
 
     /// Pointer to the next instruction in the list
     Instruction* next = nullptr;
